+++ modules/auth/controller/sendInvitation.js
@@ -6,7 +6,7 @@ const btoa = require('btoa');
 const { SCHEMA_TYPE } = require("../../../Config/schemaType.js");
 const mongoose = require("mongoose");
 const { updateCompanyFun } = require("../../company/controller/updateCompany.js");
-const { getUserByQueyFun } = require("../../usersModule/controller.js");
+const { getUserByQueyFun } = require("../../users-module/controller.js");
 const { updateMemberFunction } = require('../../settings/Members/controller.js');
 const logger = require("../../../Config/loggerConfig.js");
 const { emitListener } = require("../../company/eventController.js");
